:basketball:

+++ leetcode/algorithms/PowerOfThree.java
@@ -4,7 +4,7 @@
 public class PowerOfThree {
 
     public boolean isPowerOfThree(int n) {
-        int power = 1;
+        long power = 1;
         while (power < n) {
             power *= 3;
         }
